created separate file for taxids test
